Addressing Jessica's comments
